# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synapse Enterprise Intelligence API",
    "description": "Graphiti-backed temporal knowledge graph platform. Ingests enterprise documents, extracts a closed-vocabulary ontology, resolves entities, and maintains bi-temporal fact validity.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/ontology/entity-types/{type_id}": {
      "patch": {
        "summary": "Patch Entity Type",
        "operationId": "patch_entity_type_api_v1_ontology_entity_types__type_id__patch",
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Type Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityTypePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch Entity Type Api V1 Ontology Entity Types  Type Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ontology/changes": {
      "post": {
        "summary": "Propose Change",
        "operationId": "propose_change_api_v1_ontology_changes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeProposal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Propose Change Api V1 Ontology Changes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ontology/migrations/{change_id}/approve": {
      "post": {
        "summary": "Approve Migration",
        "operationId": "approve_migration_api_v1_ontology_migrations__change_id__approve_post",
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Change Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approve Migration Api V1 Ontology Migrations  Change Id  Approve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ontology/migrations/{change_id}/reject": {
      "post": {
        "summary": "Reject Migration",
        "operationId": "reject_migration_api_v1_ontology_migrations__change_id__reject_post",
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Change Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reject Migration Api V1 Ontology Migrations  Change Id  Reject Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/observability/consistency/{mismatch_id}/escalate": {
      "post": {
        "summary": "Escalate",
        "description": "Acknowledge a mismatch so it stops being reported.",
        "operationId": "escalate_api_v1_observability_consistency__mismatch_id__escalate_post",
        "parameters": [
          {
            "name": "mismatch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mismatch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Escalate Api V1 Observability Consistency  Mismatch Id  Escalate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeProposal": {
        "properties": {
          "relation_type": {
            "type": "string",
            "title": "Relation Type"
          },
          "field": {
            "type": "string",
            "title": "Field"
          },
          "old_value": {
            "title": "Old Value"
          },
          "new_value": {
            "title": "New Value"
          },
          "justification": {
            "type": "string",
            "title": "Justification",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "relation_type",
          "field"
        ],
        "title": "ChangeProposal",
        "description": "Body posted by `OntologyScreen.handlePropose`."
      },
...
      "EntityTypePatch": {
        "properties": {
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "EntityTypePatch"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}